favorites error redirect

+++ app/controllers/listings_controller.rb
@@ -1,5 +1,5 @@
 class ListingsController < ApplicationController
-  skip_before_action :authenticate_user!, only: [:index, :show]
+  skip_before_action :authenticate_user!, only: [:index, :show, :addfavorites]
 
   def index
     @listings = policy_scope(Listing)
@@ -65,23 +65,33 @@ class ListingsController < ApplicationController
 
   def addfavorites
     @listing = Listing.find(params[:listing_id])
-    listing_id = params[:listing_id]
-    user_id = current_user.id
 
-    fav_check = Favorite.all.find_by(listing_id: listing_id, user_id: user_id)
+    if user_signed_in?
+      listing_id = params[:listing_id]
+      user_id = current_user.id
 
-    if fav_check.nil?
-      new_fav = Favorite.new
-      new_fav.listing_id = listing_id
-      new_fav.user_id = user_id
-      new_fav.save!
+      fav_check = Favorite.all.find_by(listing_id: listing_id, user_id: user_id)
+
+      if fav_check.nil?
+        new_fav = Favorite.new
+        new_fav.listing_id = listing_id
+        new_fav.user_id = user_id
+        new_fav.save!
+      else
+        fav_check.destroy
+      end
+
+      respond_to do |format|
+        format.json { render json: {status: fav_check.nil?} }
+      end
+    
     else
-      fav_check.destroy
+      
+      respond_to do |format|
+        format.json { render json: {error: 'User not signed in. Please sign in first before adding favorites', status: "unauthorized"} }
+      end
     end
 
-    respond_to do |format|
-      format.json { render json: {status: fav_check.nil?} }
-    end
     authorize @listing
   end
